// usersweb/src/app/interfaces/User.ts
/* 
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */


export interface User {
  id: number;
  firstName: string;
  lastName: string;
  mail: string,
  idIdentityDocumentType: number,
  identification: number,
  phone: number
  
}